More Bug Fixes
Now correctly checks the 'Character' steamID (on damage) rather then the user processing the hit.

Will now get the Zones.txt file from the server, Take the numbers of X Y R and turn them into a decimal and back into a string and then to a float (This will resolve any issues with ',' to '.' in different language groups!)

Updated "NoPickup" to "NoItemPickup" in the default file.

Adding a RestrictionCheckFloatReturnCharacter & RestrictionCheckCharacter so that we can check the affected players location and adminlist check instead.

Added in some prevention of loading Util.RestrictionCheck if WorldofValheimZones.ServerMode is true (We are a server)