World of Valheim - Zones - V0.6.1 Pre-Patch
Initizliating a new permission system in the ZoneTypes.  If a user has the absolute permission node of "HackShardGaming.WoV-Zones.Override.ZONETYPE" then it will add them into the Admins list (on there local client) so that they can also override permission nodes.

Moved an OnDeath interact from StandAlonePatches into the actual OnDeath Zone-Configuration as it never gets called properly.

Added in a Chat Message string when changing zones.  This should help the user see that they changed zones even if another plugin/mod pops up a message in the center of the screen.

Added in a permission node "HackShardGaming.WoV-Zones.Wards.Override" If a user has this absolute permission node they will override the Ward Damage/Drop/Pickup even if  they are set to true on the server.

Removed our forced OnDeath RPC (which effectively removes the RespawnTimer if NoItemLoss is not set to true. in the zone / server
Added Zone Change notifications in the chat bar.  This way if something covers it up (another plugin) a user can still tell they changed zones.

Also modified ReloadZones to be more Specific per person. We now send an RPC per user connected and it includes there proper serialized information for 0.6.0 patches.